kasa_electricity_reporter_function/requirements.txt
# For Tp-link Kasa
tplink-cloud-api==1.0.2
myb-newrelic-api>=1.0.0
